==> Phylophere-integration/main_project_scripts/0.5.Compare-old-v-new.Rmd <==
---
title: "data_preparation"
author: "Miguel Ramon"
date: "`r Sys.Date()`"
output:
html_document: default
pdf_document: default
---

################################## 

# COMPARE OLD DATA V NEW DATA

### PRIMATES' NEOPLASY PROJECT

### MIGUEL RAMON ALONSO

################################## 

```{r setup, message=FALSE, warning=FALSE}

knitr::opts_knit$set(root.dir = "/home/miguel/IBE-UPF/PhD/NEOPLASY_PRIMATES")
getwd()

```

## Directory definition

```{r dirdef, message=FALSE, warning=FALSE}

workingDir <- getwd()
dataDir <- file.path(workingDir, "Data")
resultsDir <- file.path(workingDir,"Malignancy_Primates/Out/")

# Load common objects, functions and libraries
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/commons.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/primate_functions.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/primate_objects.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/dunn_modified.R"))

```

### Seed and library loading

```{r lib_load,warning=FALSE,message=FALSE}

```

### Cancer traitfile import

```{r trait_import}
# Set cancer trait dir
cancer_dir <- file.path(dataDir, "1.Cancer_data/Neoplasia_species360")

# Open traits
cancer_traits <- read.csv(file.path(cancer_dir, "cancer_traits-FINAL-DATASET-120424.csv"), sep = ",")
cancer_traits_old <- read.csv(file.path(cancer_dir, "OLD_phenotype_data_pre_120424/species360_primates_neoplasia_20230519.csv"), sep = ",")

# Remove leading/tailing whitespaces
cancer_traits[] <- lapply(cancer_traits, function(col) trimws(col))
cancer_traits_old[] <- lapply(cancer_traits_old, function(col) trimws(col))

# binarize
cancer_traits_old$label <- gsub(" ", "_", cancer_traits_old$species)

# Reorder and filter species
cancer_traits <- cancer_traits %>%
  dplyr::rename(label = species) %>%
  dplyr::rename(necropsy_count = adult_necropsy_count) %>%
  # as numeric columns from 3 onwards
  mutate(across(3:ncol(cancer_traits), as.numeric)) %>%
  dplyr::select(-common_name)

cancer_traits_old <- cancer_traits_old %>%
  dplyr::select(family,label,necropsy_count,neoplasia_necropsy,neoplasia_prevalence) %>%
  mutate(across(3:5, as.numeric)) %>%
  filter(!is.na(neoplasia_prevalence))

```

```{r data_prune}

common_names <- intersect(primates_tree$tip.label, cancer_traits$label)
common_names_old <- intersect(primates_tree$tip.label, cancer_traits_old$label)

common_names_both_datasets <- intersect(common_names, common_names_old) # Coincides

# Prune dataset
cancer_traits_clean <- cancer_traits[cancer_traits$label %in% common_names, ]
write.csv(cancer_traits_clean, "Data/1.Cancer_data/Neoplasia_species360/clean_primate_traits-labeled.csv", quote = FALSE, row.names = FALSE, col.names = TRUE)
cancer_traits_old_clean <- cancer_traits_old[cancer_traits_old$label %in% common_names_old, ]
write.csv(cancer_traits_old_clean, "Data/1.Cancer_data/Neoplasia_species360/OLD_phenotype_data_pre_120424/clean_primate_traits_OLD.csv", quote = FALSE, row.names = FALSE, col.names = TRUE)

```

# Load melted stuff

```{r load_melt}
# Define path to melted traits file
melt_file <- file.path(resultsDir, "1.Data-exploration/1.Species_distribution/melted_traits.csv")

# Load melted traits data
melted_traits <- read.csv(melt_file)

melt_old <- read.csv("Out/0.5.Dataset-comparison/old_melted_traits.csv")
```

# Start comparing things

```{r compare_things, fig.height=7, fig.width=10, dev='png'}
# Compare species number and in-family distributions

common_species <- intersect(cancer_traits_old_clean$label, cancer_traits_clean$label)
added_species <- setdiff(cancer_traits_clean$label, cancer_traits_old_clean$label)
deleted_species <- setdiff(cancer_traits_old_clean$label, cancer_traits_clean$label)

family_distribution <- cancer_traits_clean %>%
  group_by(family) %>%
  summarise(count = n())

# Count families in both datasets
old_family_counts <- cancer_traits_old_clean %>% count(family)
new_family_counts <- cancer_traits_clean %>% count(family)

# Merge counts into one dataframe
family_comparison <- merge(old_family_counts, new_family_counts, by = "family", all = TRUE)
colnames(family_comparison) <- c("Family", "Old_Count", "New_Count")
family_comparison[is.na(family_comparison)] <- 0

print (family_comparison)

# Reshape data for plotting
family_comparison_long <- family_comparison %>%
  pivot_longer(cols = c(Old_Count, New_Count), names_to = "Dataset", values_to = "Count")

# Plotting
compare_plottie <- ggplot(family_comparison_long, aes(x = Family, y = Count, fill = Dataset)) +
  geom_bar(stat = "identity", position = "dodge") +
  geom_text(aes(label=Count), vjust=1.6, color="black", fontface = "bold",
      position = position_dodge(0.9), size=6) +
  labs(x = "Family", y = "Count", title = "Comparison of Family Distribution") +
  scale_fill_manual(values = c("Old_Count" = "skyblue", "New_Count" = "salmon"), labels = c("Old_Count" = "Family old data", "New_Count" = "Family new data")) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1, size = 12),
    plot.title = element_text(size = 16, face = "bold", color = "black", hjust = 0.5, margin = margin(t = 20, b = 15)),
    plot.subtitle = element_text(size = 13, face = "bold", color = "black", hjust = 0.5, margin = margin(t = 20, b = 10)),
    plot.caption = element_text(size = 13, face = "bold", margin = margin(t = 10, b = 20)),
    axis.title.x = element_text(size = 13, hjust = 0.5, margin = margin(t = 20, b = 20)),
    axis.title.y = element_text(size = 13, angle = 90, hjust = 0.5, margin = margin(l = 20, r = 20)))
compare_plottie

# Family plot
family_plot <- ggplot(family_distribution, aes(x = family, y = count, fill = family)) +
  geom_bar(stat = "identity") +
  geom_text(aes(label=count), vjust=0.5, hjust = -1.5, color="black", fontface = "bold", size=6) +
  labs(x = "Family", y = "Count", title = "Family Distribution") +
  theme_minimal() +
  scale_fill_manual(values = primate_family_colors, breaks = 0) +
  # Increase y axis up to 15
  scale_y_continuous(limits = c(0, 15), expand = c(0, 0)) +
  theme(
    axis.text.y = element_text(hjust = 1, size = 18, face = "bold", color = "black"),
    axis.text.x = element_blank(),
    plot.title = element_blank(),
    axis.title.x = element_text(size = 18, hjust = 0.5, margin = margin(t = 20), face = "bold", color = "black"),
    axis.title.y = element_blank(),
    legend.position = "none"
  )
# Swap plot axis
family_plot <- family_plot + coord_flip()
family_plot




# Save the plot
compare_plot_path <- file.path("Out/0.5.Dataset-comparison/family_comparison")
createDir(compare_plot_path)
ggsave(file.path(compare_plot_path, "family_comparison.png"), compare_plottie, device = "png", width = 10, height = 7, dpi = 300)
ggsave(file.path(compare_plot_path, "family_distribution.png"), family_plot, device = "png", width = 12, height = 7, dpi = "retina")

# Boxplot explaining significant differences between the traits
new_traits <- unique(melt_new$Trait) # Get unique traits

```
