==> Phylophere-integration/main_project_scripts/5.3.CAAS_analysis.Rmd <==
---
title: "CAAS analysis"
author: "Miguel Ramon"
date: "`r Sys.Date()`"
citations:   
output:
  html_document: default
  pdf_document: default
---

# CAAS analysis
  
This document is used to generate the results for the CAAS analysis. The analysis is divided into several sections, each of which is detailed below. The analysis is performed on the data generated by the CAAS pipeline, which is stored in the `2.CAAS/1.Discovery/2.Results` folder. The results are stored in the `2.CAAS/1.Discovery/3.5.CAAS_exploration` folder.

## Setup and Directory Configuration
This section configures the working environment, sets directories, and loads necessary functions and libraries.

```{r setup, message=FALSE, warning=FALSE}

# Suppress warnings and messages in output
knitr::opts_chunk$set(warning = FALSE, message = FALSE)
# Set the root directory for the project
knitr::opts_knit$set(root.dir = "/home/miguel/IBE-UPF/PhD/NEOPLASY_PRIMATES")
# Display current working directory
getwd()
# Set seed for reproducibility
set.seed(1998)

```

```{r dirdef, message=FALSE, warning=FALSE}

# Define working, data, and results directories
workingDir <- getwd()
dataDir <- file.path(workingDir, "Data")
resultsDir <- file.path(workingDir, "Malignancy_Primates/Out")

# Load libraries, common functions, primate-specific functions and objects
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/libraries.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/commons.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/primate_functions.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/primate_objects.R"))

# Note that all variables may be changed from COMMONS and PRIMATE_OBJECTS scripts.

```

## 1. P-value nomination
In this section we will filter the CAAS results by p-value and pattern. We will create a list of data frames for each scenario and pattern, and save the filtered data frames to files.

```{r p-value, message=FALSE, warning=FALSE}

caas_folder <- file.path(resultsDir, "2.CAAS/3.CAAS_postproc/4.Declustered_i/")

# From caas_folder/$TRAIT/Discovery, extract files
caas_files <- list.files(path = caas_folder, pattern = "\\.tsv$", full.names = TRUE, recursive = TRUE)

# Read files and split dataframes
setup_dfs <- lapply(caas_files, read_tsv)
names(setup_dfs) <- gsub("_sp.tab.caas_declustered.tsv", "", basename(caas_files))

setup_dfs.05 <- list()  # New list for filtered data frames

# Function to filter rows and create id2 column

filter_gap <- function(df) {
  df %>%
    filter(GFG == 0 & GBG == 0)
}

setup_dfs <- lapply(setup_dfs, filter_gap)

filter_and_mutate <- function(df, name) {
  df <- df %>%
    # Filter by pvalue
    filter(as.numeric(Pvalue) <= 0.05)
}

# Apply the filtering and mutation logic to each sublist in setup_dfs
setup_dfs.05 <- lapply(names(setup_dfs), function(name) filter_and_mutate(setup_dfs[[name]], name))
names(setup_dfs.05) <- names(setup_dfs)

filter_and_mutate_pattern <- function(df, name, pattern) {
  df <- df %>%
    filter(Pattern == pattern)
}

# Extract the filtered lists
extract_lists <- function(df, name, filt_type){
  # Extract first bit from name for trait_name
  trait_name <- strsplit(name, "_")[[1]][1]
  # Extract second bit for file_name
  file_name <- paste(strsplit(name, "_")[[1]][3], strsplit(name, "_")[[1]][4], sep = "_")
  path <- file.path(paste0(caas_list_dir,"/1.CAAS_tables_filtered/", trait_name, "/", file_name)) # add /pattern4/ for pattern4
  createDir(path)
  write.table(df, file.path(path, paste0(filt_type, "_", file_name, ".output")), quote = FALSE, row.names = FALSE, col.names = TRUE, sep = "\t")

  return(df)
}

# Set filter list out
caas_list_dir <- file.path(resultsDir, "2.CAAS/3.5.CAAS_exploration")
createDir(caas_list_dir)

# Apply the filtering and mutation logic to each sublist in setup_dfs
# Apply filter_and_mutate_pattern 
pat_1_dfs <- lapply(names(setup_dfs), function(name) filter_and_mutate_pattern(setup_dfs[[name]], name, "pattern1"))
pat_2_dfs <- lapply(names(setup_dfs), function(name) filter_and_mutate_pattern(setup_dfs[[name]], name, "pattern2"))
pat_3_dfs <- lapply(names(setup_dfs), function(name) filter_and_mutate_pattern(setup_dfs[[name]], name, "pattern3"))
names(pat_1_dfs) <- names(setup_dfs)
names(pat_2_dfs) <- names(setup_dfs)
names(pat_3_dfs) <- names(setup_dfs)

# Create pat_12 and pat_13 by merging pat_1 and pat_2, and pat_1 and pat_3
pat_12_dfs <- lapply(names(pat_1_dfs), function(name) bind_rows(pat_1_dfs[[name]], pat_2_dfs[[name]]))
names(pat_12_dfs) <- names(pat_1_dfs)
pat_13_dfs <- lapply(names(pat_1_dfs), function(name) bind_rows(pat_1_dfs[[name]], pat_3_dfs[[name]]))
names(pat_13_dfs) <- names(pat_1_dfs)


# Adhoc modify function structure to introduce pattern4 location
caas_files.pattern1 <- lapply(names(pat_1_dfs), function(name) extract_lists(pat_1_dfs[[name]], name, "pat1_05"))
names(caas_files.pattern1) <- names(setup_dfs.05)
caas_files.pattern2 <- lapply(names(pat_2_dfs), function(name) extract_lists(pat_2_dfs[[name]], name, "pat2_05"))
names(caas_files.pattern2) <- names(setup_dfs.05)
caas_files.pattern3 <- lapply(names(pat_3_dfs), function(name) extract_lists(pat_3_dfs[[name]], name, "pat3_05"))
names(caas_files.pattern3) <- names(setup_dfs.05)
caas_files.pattern12 <- lapply(names(pat_12_dfs), function(name) extract_lists(pat_12_dfs[[name]], name, "pat12_05"))
names(caas_files.pattern12) <- names(setup_dfs.05)
caas_files.pattern13 <- lapply(names(pat_13_dfs), function(name) extract_lists(pat_13_dfs[[name]], name, "pat13_05"))
names(caas_files.pattern13) <- names(setup_dfs.05)

# Now lets do the same with the significant CAAS
# Apply filter_and_mutate_pattern 
pat_1_dfs.05 <- lapply(names(setup_dfs.05), function(name) filter_and_mutate_pattern(setup_dfs.05[[name]], name, "pattern1"))
pat_2_dfs.05 <- lapply(names(setup_dfs.05), function(name) filter_and_mutate_pattern(setup_dfs.05[[name]], name, "pattern2"))
pat_3_dfs.05 <- lapply(names(setup_dfs.05), function(name) filter_and_mutate_pattern(setup_dfs.05[[name]], name, "pattern3"))
names(pat_1_dfs.05) <- names(setup_dfs.05)
names(pat_2_dfs.05) <- names(setup_dfs.05)
names(pat_3_dfs.05) <- names(setup_dfs.05)

# Create pat_12 and pat_13 by merging pat_1 and pat_2, and pat_1 and pat_3
pat_12_dfs.05 <- lapply(names(pat_1_dfs.05), function(name) bind_rows(pat_1_dfs.05[[name]], pat_2_dfs.05[[name]]))
names(pat_12_dfs.05) <- names(pat_1_dfs.05)
pat_13_dfs.05 <- lapply(names(pat_1_dfs.05), function(name) bind_rows(pat_1_dfs.05[[name]], pat_3_dfs.05[[name]]))
names(pat_13_dfs.05) <- names(pat_1_dfs.05)

# Adhoc modify function structure to introduce pattern4 location
caas_files.pattern1.05 <- lapply(names(pat_1_dfs.05), function(name) extract_lists(pat_1_dfs.05[[name]], name, "pat1_05"))
names(caas_files.pattern1.05) <- names(setup_dfs.05)
caas_files.pattern2.05 <- lapply(names(pat_2_dfs.05), function(name) extract_lists(pat_2_dfs.05[[name]], name, "pat2_05"))
names(caas_files.pattern2.05) <- names(setup_dfs.05)
caas_files.pattern3.05 <- lapply(names(pat_3_dfs.05), function(name) extract_lists(pat_3_dfs.05[[name]], name, "pat3_05"))
names(caas_files.pattern3.05) <- names(setup_dfs.05)
caas_files.pattern12.05 <- lapply(names(pat_12_dfs.05), function(name) extract_lists(pat_12_dfs.05[[name]], name, "pat12_05"))
names(caas_files.pattern12.05) <- names(setup_dfs.05)
caas_files.pattern13.05 <- lapply(names(pat_13_dfs.05), function(name) extract_lists(pat_13_dfs.05[[name]], name, "pat13_05"))
names(caas_files.pattern13.05) <- names(setup_dfs.05)

```

## 2. Permulation p-value nomination
And now what we will do is to load the permuted data and see how many CAAS stay significant. We will do it using both the nominally significant and the non-nominally significant CAAS

```{r bootstrap data}

boot_folder <- file.path(resultsDir, "2.CAAS/2.Results")

# From caas_folder/$TRAIT/Discovery, extract files
boot_files <- list.files(path = boot_folder, pattern = "\\.tab.caas$", full.names = TRUE, recursive = TRUE)

# Now we load the boostrap data to see how many CAAS stay significant. We will do it using both the nominally significant and the non-nominally significant CAAS
caas_files_boot <- boot_files[grepl("/bootstrap/", boot_files, ignore.case = TRUE)]

# Read files and split dataframes
boot_dfs <- lapply(caas_files_boot, read_tsv, col_names = FALSE)
names(boot_dfs) <- gsub("_sp.tab.caas", "", basename(caas_files_boot))

boot_dfs.05 <- list()  # New list for filtered data frames

filter_and_mutate <- function(df, name) {
  df <- data.frame(df)  %>%
    # 0 p-values are considered significant but they don't make sense from a biological point of view
    mutate(X4 = ifelse(X2 == 0, 1/(X3 + 1), X4)) %>%
    # Drop the last column
    dplyr::rename(GenePos = X1, Counts = X2, Perms = X3, Pvalue.boot = X4) %>%
    # Correct Pvalue.boot with BH correction
    mutate(Pvalue.boot.BH = p.adjust(Pvalue.boot, method = "BH"))  %>%
    filter(as.numeric(Pvalue.boot) <= 0.05 | as.numeric(Pvalue.boot) == 0)
}

# Apply the filtering and mutation logic to each sublist in setup_dfs
boot_dfs.05 <- lapply(names(boot_dfs), function(name) filter_and_mutate(boot_dfs[[name]], name))

# # Save rds
# saveRDS(boot_dfs, file = file.path(boot_folder, "boot_dfs05.rds"))
# boot_dfs.05 <- readRDS(file = file.path(boot_folder, "boot_dfs05.rds"))
names(boot_dfs.05) <- names(setup_dfs)

# Create GenePos column in setup_dfs and setup_dfs.05
setup_dfs <- lapply(setup_dfs, function(df) mutate(df, GenePos = paste0(Gene, "@", Position)))
setup_dfs.05 <- lapply(setup_dfs.05, function(df) mutate(df, GenePos = paste0(Gene, "@", Position)))

# Filter setup_dfs and setup_dfs.05 by boot_dfs.05
filter_boot <- function(df, boot_df) {
  df <- df %>%
    # Filter by GenePos
    filter(GenePos %in% boot_df$GenePos) %>%
    # Add Pvalue.boot column
    left_join(boot_df %>% select(GenePos, Pvalue.boot), by = "GenePos")
}

setup_dfs_boot <- lapply(names(setup_dfs), function(name) filter_boot(setup_dfs[[name]], boot_dfs.05[[name]]))
names(setup_dfs_boot) <- names(setup_dfs)
setup_dfs.05_boot <- lapply(names(setup_dfs.05), function(name) filter_boot(setup_dfs.05[[name]], boot_dfs.05[[name]]))
names(setup_dfs.05_boot) <- names(setup_dfs.05)

# Extract the lists
extract_lists <- function(df, name, filt_type){
  # Extract first bit from name for trait_name
  trait_name <- strsplit(name, "_")[[1]][1]
  # Extract second bit for file_name
  file_name <- paste(strsplit(name, "_")[[1]][3], strsplit(name, "_")[[1]][4], sep = "_")
  path <- file.path(paste0(caas_list_dir,"/1.CAAS_tables_filtered/", trait_name, "/", file_name, "/boot")) # add /pattern4/ for pattern4
  createDir(path)
  write.table(df, file.path(path, paste0(filt_type, "_", file_name, ".output")), quote = FALSE, row.names = FALSE, col.names = TRUE, sep = "\t")

  return(df)
}

# Save the filtered data frames
caas_files_boot <- lapply(names(setup_dfs_boot), function(name) extract_lists(setup_dfs_boot[[name]], name, "boot"))
names(caas_files_boot) <- names(setup_dfs_boot)

caas_files_boot.05 <- lapply(names(setup_dfs.05_boot), function(name) extract_lists(setup_dfs.05_boot[[name]], name, "pval05_boot"))
names(caas_files_boot.05) <- names(setup_dfs.05_boot)

# Filter the patterns
filter_and_mutate_pattern <- function(df, name, pattern) {
  df <- df %>%
    filter(Pattern == pattern)
}

pat_1_dfs.boot <- lapply(names(caas_files_boot), function(name) filter_and_mutate_pattern(caas_files_boot[[name]], name, "pattern1"))
names(pat_1_dfs.boot) <- names(caas_files_boot)
pat_2_dfs.boot <- lapply(names(caas_files_boot), function(name) filter_and_mutate_pattern(caas_files_boot[[name]], name, "pattern2"))
names(pat_2_dfs.boot) <- names(caas_files_boot)
pat_3_dfs.boot <- lapply(names(caas_files_boot), function(name) filter_and_mutate_pattern(caas_files_boot[[name]], name, "pattern3"))
names(pat_3_dfs.boot) <- names(caas_files_boot)
pat_12_dfs.boot <- lapply(names(caas_files_boot), function(name) bind_rows(pat_1_dfs.boot[[name]], pat_2_dfs.boot[[name]]))
names(pat_12_dfs.boot) <- names(caas_files_boot)
pat_13_dfs.boot <- lapply(names(caas_files_boot), function(name) bind_rows(pat_1_dfs.boot[[name]], pat_3_dfs.boot[[name]]))
names(pat_13_dfs.boot) <- names(caas_files_boot)

# Save the filtered data frames
caas_files_pattern1_boot <- lapply(names(pat_1_dfs.boot), function(name) extract_lists(pat_1_dfs.boot[[name]], name, "pat1_boot"))
names(caas_files_pattern1_boot) <- names(pat_1_dfs.boot)
caas_files_pattern2_boot <- lapply(names(pat_2_dfs.boot), function(name) extract_lists(pat_2_dfs.boot[[name]], name, "pat2_boot"))
names(caas_files_pattern2_boot) <- names(pat_2_dfs.boot)
caas_files_pattern3_boot <- lapply(names(pat_3_dfs.boot), function(name) extract_lists(pat_3_dfs.boot[[name]], name, "pat3_boot"))
names(caas_files_pattern3_boot) <- names(pat_3_dfs.boot)
caas_files_pattern12_boot <- lapply(names(pat_12_dfs.boot), function(name) extract_lists(pat_12_dfs.boot[[name]], name, "pat12_boot"))
names(caas_files_pattern12_boot) <- names(pat_12_dfs.boot)
caas_files_pattern13_boot <- lapply(names(pat_13_dfs.boot), function(name) extract_lists(pat_13_dfs.boot[[name]], name, "pat13_boot"))
names(caas_files_pattern13_boot) <- names(pat_13_dfs.boot)

pat_1_dfs.boot.05 <- lapply(names(caas_files_boot.05), function(name) filter_and_mutate_pattern(caas_files_boot.05[[name]], name, "pattern1"))
names(pat_1_dfs.boot.05) <- names(caas_files_boot.05)
pat_2_dfs.boot.05 <- lapply(names(caas_files_boot.05), function(name) filter_and_mutate_pattern(caas_files_boot.05[[name]], name, "pattern2"))
names(pat_2_dfs.boot.05) <- names(caas_files_boot.05)
pat_3_dfs.boot.05 <- lapply(names(caas_files_boot.05), function(name) filter_and_mutate_pattern(caas_files_boot.05[[name]], name, "pattern3"))
names(pat_3_dfs.boot.05) <- names(caas_files_boot.05)
pat_12_dfs.boot.05 <- lapply(names(caas_files_boot.05), function(name) bind_rows(pat_1_dfs.boot.05[[name]], pat_2_dfs.boot.05[[name]]))
names(pat_12_dfs.boot.05) <- names(caas_files_boot.05)
pat_13_dfs.boot.05 <- lapply(names(caas_files_boot.05), function(name) bind_rows(pat_1_dfs.boot.05[[name]], pat_3_dfs.boot.05[[name]]))
names(pat_13_dfs.boot.05) <- names(caas_files_boot.05)

# Save the filtered data frames
caas_files_pattern1_boot.05 <- lapply(names(pat_1_dfs.boot.05), function(name) extract_lists(pat_1_dfs.boot.05[[name]], name, "pat1_boot05"))
names(caas_files_pattern1_boot.05) <- names(pat_1_dfs.boot.05)
caas_files_pattern2_boot.05 <- lapply(names(pat_2_dfs.boot.05), function(name) extract_lists(pat_2_dfs.boot.05[[name]], name, "pat2_boot05"))
names(caas_files_pattern2_boot.05) <- names(pat_2_dfs.boot.05)
caas_files_pattern3_boot.05 <- lapply(names(pat_3_dfs.boot.05), function(name) extract_lists(pat_3_dfs.boot.05[[name]], name, "pat3_boot05"))
names(caas_files_pattern3_boot.05) <- names(pat_3_dfs.boot.05)
caas_files_pattern12_boot.05 <- lapply(names(pat_12_dfs.boot.05), function(name) extract_lists(pat_12_dfs.boot.05[[name]], name, "pat12_boot05"))
names(caas_files_pattern12_boot.05) <- names(pat_12_dfs.boot.05)
caas_files_pattern13_boot.05 <- lapply(names(pat_13_dfs.boot.05), function(name) extract_lists(pat_13_dfs.boot.05[[name]], name, "pat13_boot05"))
names(caas_files_pattern13_boot.05) <- names(pat_13_dfs.boot.05)

```

## 3. Summarize the data
After filtering the data, we will summarize the results by counting the number of significant CAAS for each trait and pattern. We will create a summary table that includes the number of CAAS for each trait and pattern, as well as the number of CAAS that are significant in both the nominal and permuted data.

```{r summarize_data, fig.width=16, fig.height=12, message=FALSE, warning=FALSE, dpi = 320, dev="png"}

# Create a summary table for the raw and filtered dataframes
summary_table <- data.frame(
  File = names(setup_dfs),
  All = sapply(setup_dfs, nrow),
  Perm.05 = sapply(setup_dfs_boot, nrow),
  Hyp.05 = sapply(setup_dfs.05, nrow),
  Perm_Hyp.05 = sapply(setup_dfs.05_boot, nrow),
  Pattern1_All = sapply(pat_1_dfs, nrow),  
  Pattern1_Perm.05 = sapply(pat_1_dfs.boot, nrow),
  Pattern1_Hyp.05 = sapply(pat_1_dfs.05, nrow),
  Pattern1_Perm_Hyp.05 = sapply(pat_1_dfs.boot.05, nrow),
  Pattern2_All = sapply(pat_2_dfs, nrow),
  Pattern2_Perm.05 = sapply(pat_2_dfs.boot, nrow),
  Pattern2_Hyp.05 = sapply(pat_2_dfs.05, nrow),
  Pattern2_Perm_Hyp.05 = sapply(pat_2_dfs.boot.05, nrow),
  Pattern3_All = sapply(pat_3_dfs, nrow),
  Pattern3_Perm.05 = sapply(pat_3_dfs.boot, nrow),
  Pattern3_Hyp.05 = sapply(pat_3_dfs.05, nrow),
  Pattern3_Perm_Hyp.05 = sapply(pat_3_dfs.boot.05, nrow),
  Pattern12_All = sapply(pat_12_dfs, nrow),
  Pattern12_Perm.05 = sapply(pat_12_dfs.boot, nrow),
  Pattern12_Hyp.05 = sapply(pat_12_dfs.05, nrow),
  Pattern12_Perm_Hyp.05 = sapply(pat_12_dfs.boot.05, nrow),
  Pattern13_All = sapply(pat_13_dfs, nrow),
  Pattern13_Perm.05 = sapply(pat_13_dfs.boot, nrow),
  Pattern13_Hyp.05 = sapply(pat_13_dfs.05, nrow),
  Pattern13_Perm_Hyp.05 = sapply(pat_13_dfs.boot.05, nrow)
)

totals <- c("Total", colSums(summary_table[,-1]))

# Read traitfiles explicit on which species were used
malignant_pairs.f <- file.path(resultsDir, "2.CAAS/1.Discovery/1.Traitfiles/malignant_prevalence/malignant_prevalence_paired_species.tab")
malignant_pairs <- read_tsv(malignant_pairs.f, col_names = TRUE) %>%
  dplyr::select(species, pair, contrast) %>%
  unique() %>%
  group_by(pair, contrast) %>%
  summarize(sp_pair = paste(species, collapse = ",")) %>%
  ungroup() %>%
  group_by(contrast) %>%
  summarize(sp_pair.d = paste(sp_pair, collapse = "\n")) %>%
  ungroup()

neoplasia_pairs.f <- file.path(resultsDir, "2.CAAS/1.Discovery/1.Traitfiles/neoplasia_prevalence/neoplasia_prevalence_paired_species.tab")
neoplasia_pairs <- read_tsv(neoplasia_pairs.f, col_names = TRUE) %>%  
  dplyr::select(species, pair, contrast) %>%
  unique() %>%
  group_by(pair, contrast) %>%
  summarize(sp_pair = paste(species, collapse = ",")) %>%
  ungroup() %>%
  group_by(contrast) %>%
  summarize(sp_pair.d = paste(sp_pair, collapse = "\n")) %>%
  ungroup()


# Optionally, if you want overall totals (summing over all files):
summary_table <- rbind(summary_table, totals) %>%
  # Remove row names
  rownames_to_column(var = "File2") %>%
  select(-File2)

# Create a combined lookup table for species pairs by contrast
lookup_table <- bind_rows(
  malignant_pairs %>% mutate(source = "malignant_pairs"),
  neoplasia_pairs %>% mutate(source = "neoplasia_pairs")
) %>%
  mutate(file_name = paste0(source, "_", contrast))

# Update summary_table by matching File2 to the lookup table
summary_table2 <- summary_table %>%
  mutate(
    contrast_num = str_extract(File, "\\d+"),  # Extract contrast number
    source = case_when(
      str_detect(File, "^malignant_prevalence_pair_") ~ "malignant_pairs",
      str_detect(File, "^neoplasia_prevalence_pair_") ~ "neoplasia_pairs",
      TRUE ~ NA_character_
    ),
    file_name = paste0(source, "_", contrast_num)
  ) %>%
  left_join(
    select(lookup_table, file_name, sp_pair.d),
    by = "file_name"
  ) %>%
  mutate(
    File = case_when(
      !is.na(source) ~ paste0(source, "_", contrast_num),
      TRUE ~ File
    )
  ) %>%
  dplyr::select(sp_pair.d, everything())

# Print the summary table
print(summary_table2)

# Write the summary table to a file
write.table(summary_table2, file = file.path(caas_list_dir, "summary_table.txt"), row.names = FALSE, col.names = TRUE, quote = FALSE, sep = "\t")

# Assuming summary_plot_data is already loaded and excludes "Total" as in your code
summary_plot_data <- summary_table2 %>%
  drop_na(sp_pair.d)

# Ensure Category is a factor with the specified order
# Assuming sp_pair.d is already in the desired order; if not, specify levels explicitly
summary_plot_data2 <- summary_plot_data %>%
  mutate(sp_pair.d = factor(sp_pair.d, levels = unique(sp_pair.d)))

# Define color palette
levels_palette <- c("All" = "skyblue3", 
                    "Hyp.05" = "salmon3", 
                    "Perm.05" = "darkseagreen3", 
                    "Perm_Hyp.05" = "tan1")

# Prepare overall data
overall_data <- summary_plot_data[, c("sp_pair.d", "source", "All", "Hyp.05", "Perm.05", "Perm_Hyp.05")]

# Reshape to long format
overall_long <- reshape2::melt(overall_data, id.vars = c("sp_pair.d", "source"),
                               variable.name = "Category", value.name = "Count") %>%
  mutate(Count = as.numeric(Count))

# Create the dodged bar plot
p_overall <- ggplot(overall_long, aes(x = sp_pair.d, y = Count, fill = Category)) +
  geom_bar(stat = "identity", position = "dodge") +
  facet_grid(~ source, scales = "free_x") +  # Use free_x to show only relevant x-axis labels
  theme_minimal() +
  labs(title = "Overall: Raw and Bootstrap Counts", x = "Species Pair", y = "Count") +
  scale_fill_manual(values = levels_palette) +
  ylim(limits = c(0, 11000)) +  # Adjust y-axis limit for better visibility
  # Add more granular y-axis ticks
  scale_y_continuous(breaks = seq(0, max(overall_long$Count), by = 500)) +
  theme(axis.text.x = element_text(angle = 60, hjust = 1, vjust = 1, size = 10),
        # Include internal lines
        panel.grid.major = element_line(color = "darkgray", size = 0.5),
        panel.grid.minor = element_line(color = "lightgray", size = 0.5),
        # Set axis title margins
        axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0)),
        axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)))

# Display the plot
print(p_overall)

# Assuming summary_plot_data is the data frame with the relevant columns
# Define patterns and categories
patterns <- c("1", "2", "3")
categories <- c("All", "Hyp.05", "Perm.05", "Perm_Hyp.05")

# Initialize a list to store data frames for each category
data_list <- list()

# Loop through each category to prepare the data
for (cat in categories) {
  suffix <- paste0("_", cat)
  prefix <- "Pattern"
  
  # Construct pattern-specific column names
  pattern_cols <- paste0(prefix, patterns, suffix)
  
  print(pattern_cols)

  # Select sp_pair.d, source and pattern columns from the data
  temp_data <- summary_plot_data[, c("sp_pair.d", "source", pattern_cols)]

  # Melt to long format
  temp_long <- reshape2::melt(temp_data, id.vars = c("sp_pair.d","source"),
                              variable.name = "Pattern",
                              value.name = "Count")

  # Extract the pattern identifier (e.g., "1", "12") by removing prefix and suffix
  temp_long$Pattern <- gsub(paste0("^", prefix), "", temp_long$Pattern)
  temp_long$Pattern <- gsub(paste0(suffix, "$"), "", temp_long$Pattern)

  # Add the category column
  temp_long$Category <- cat

  # Ensure Count is numeric
  temp_long$Count <- as.numeric(temp_long$Count)

  # Store in the list
  data_list[[cat]] <- temp_long
}

# Combine all data frames into one
all_data_long <- do.call(rbind, data_list)

# Define color palette
levels_palette <- c("1" = "skyblue3", 
                    "2" = "salmon3", 
                    "3" = "darkseagreen3")

# Ensure Category is a factor with the specified order
all_data_long <- all_data_long %>%
  mutate(Category = factor(Category, levels = c("All", "Hyp.05", "Perm.05", "Perm_Hyp.05"))) %>%
  mutate(sp_pair.d = factor(sp_pair.d, levels = unique(sp_pair.d)))

# Create the stacked bar plot
p_stacked <- ggplot(all_data_long, aes(x = sp_pair.d, y = Count, fill = Pattern)) +
  geom_bar(stat = "identity") +  # Stacks bars by pattern
  facet_grid(Category ~ source, scales = "free") +  # Separate panel for each category
  theme_minimal() +
  labs(title = "Stacked Counts by Pattern for Each Category and Hypothesis",
       x = "Hypothesis", y = "Count", fill = "Pattern") +
  scale_fill_manual(values = levels_palette) +  
  # ylim(limits = c(0, 10000)) +  # Adjust y-axis limit for better visibility
  theme(axis.text.x = element_text(angle = 60, hjust = 1, vjust = 1, size = 10),
        # Include internal lines
        panel.grid.major = element_line(color = "darkgray", size = 0.5),
        panel.grid.minor = element_line(color = "lightgray", size = 0.5),
        # Set axis title margins
        axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0)),
        axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)))

# Display the plot
print(p_stacked)


# Save the two plots
ggsave(file.path(caas_list_dir, "overall_counts.png"), plot = p_overall, width = 15, height = 10, dpi = 320)
ggsave(file.path(caas_list_dir, "stacked_counts.png"), plot = p_stacked, width = 10, height = 8, dpi = 320)

```

## 4. CAAS counting
Here we will count the number of CAAS for each pattern and FFGN category. We will also save the counts to files.

```{r caas_counts, message=FALSE, warning=FALSE}

FFGN_values <- c("2", "3", "4", "5", "6", "7", "8", "9")
pattern_values <- c("pattern1", "pattern2", "pattern3")

generate_counts_df <- function(df, category, values, key) {
  counts <- sapply(values, function(value) {
    sum(df[[category]] == value)
  })
  counts_df <- data.frame(
    Category = category,
    Value = values,
    Counts = counts,
    Key = key
  )

  return(counts_df)
}

# Function to process each sublist and generate counts data frames
process_scenario <- function(scenario_df) {
  # Generate counts data frames for different categories
  
  scenario_ffgn <- generate_counts_df(scenario_df, "FFGN", FFGN_values, "FFGN")
  pattern_counts <- generate_counts_df(scenario_df, "Pattern", pattern_values, "Pattern")
  
  # Combine all counts data frames
  counts_combined <- rbind(scenario_ffgn, pattern_counts)
  return(counts_combined)
}

# Apply the process_scenario function to each sublist in setup_dfs.05
data_counts <- lapply(setup_dfs.05, process_scenario)
# data_counts_pattern4 <- lapply(pat_4_dfs, process_scenario)

# Set count out dir


def_counts <- function(sublist, sublist_name, key) {
  sublist <- sublist %>%
    dplyr::filter(Counts != 0) %>%
    dplyr::select(Value, Counts) %>%
    pivot_wider(names_from = Value, values_from = Counts)
  
  merged_df <- bind_rows(sublist)
  countDir <- file.path(caas_list_dir, "2.Species-v-counts", key)
  createDir(countDir)
  # Write to file
  write.table(merged_df, file.path(countDir, sublist_name), quote = FALSE, row.names = FALSE, col.names = TRUE, sep = "\t")
  
  return(merged_df)
}

merged_data <- lapply(names(data_counts), function(name) def_counts(data_counts[[name]], name, "pval05"))
# merged_data_pattern4 <- lapply(names(data_counts_pattern4), function(name) def_counts(data_counts_pattern4[[name]], name, "pattern4"))

names(merged_data) <- names(setup_dfs.05)

```

## 5. Gene tables
In this section, we will extract the gene information from the CAAS data and save it to files. We will create gene tables for both the nominally significant and non-nominally significant CAAS, as well as for the bootstrap data.

```{r gene_extraction, message=FALSE, warning=FALSE}
# Gene table paths
gene_path <- file.path(caas_list_dir, "3.Gene-tables")
createDir(gene_path)

# Generate gene tables
get_gene_info <- function(data, name, criteria) {
  # Extract everything up to second _ from name
  trait_name <- strsplit(name, "_")[[1]][1:2] %>% paste(collapse = "_")
  
  # Extract everything up to from second _ to . from name
  scenario_name <- strsplit(name, "_")[[1]][3:4] %>% paste(collapse = "_")
  scenario_name <- gsub("\\.tab\\.caas", "", scenario_name)
  
  path <- file.path(gene_path, trait_name, scenario_name) # add /pattern4/ for pattern4
  createDir(path)

  genes <- data$Gene[data$FFGN %in% FFGN_values]

  # Write to file
  unique_genes <- unique(genes)

    # Write gene list to file
  write.table(unique_genes, file.path(path, paste0(criteria, "_", name, ".genes")), quote = FALSE, row.names = FALSE, col.names = FALSE, sep = "\n")

  return(genes)
}
gene_tables <- lapply(names(setup_dfs), function(name) get_gene_info(setup_dfs[[name]], name, "non_sig"))
names(gene_tables) <- names(setup_dfs)
gene_tables.05 <- lapply(names(setup_dfs.05), function(name) get_gene_info(setup_dfs[[name]], name, "non_sig"))
names(gene_tables.05) <- names(setup_dfs.05)
# Do the same for pat_1_dfs, pat_2_dfs, pat_3_dfs, pat_12_dfs, pat_13_dfs
pat_1_gene_tables <- lapply(names(pat_1_dfs), function(name) get_gene_info(pat_1_dfs[[name]], name, "pat1"))
names(pat_1_gene_tables) <- names(pat_1_dfs)
pat_2_gene_tables <- lapply(names(pat_2_dfs), function(name) get_gene_info(pat_2_dfs[[name]], name, "pat2"))
names(pat_2_gene_tables) <- names(pat_2_dfs)
pat_3_gene_tables <- lapply(names(pat_3_dfs), function(name) get_gene_info(pat_3_dfs[[name]], name, "pat3"))
names(pat_3_gene_tables) <- names(pat_3_dfs)
pat_12_gene_tables <- lapply(names(pat_12_dfs), function(name) get_gene_info(pat_12_dfs[[name]], name, "pat12"))
names(pat_12_gene_tables) <- names(pat_12_dfs)
pat_13_gene_tables <- lapply(names(pat_13_dfs), function(name) get_gene_info(pat_13_dfs[[name]], name, "pat13"))
names(pat_13_gene_tables) <- names(pat_13_dfs)

# Merge all gene tables into a single data frame
all_gene_tables <- function(gene_tables) {
  
  trait_name <- strsplit(names(gene_tables)[1], "_")[[1]][1:2] %>% paste(collapse = "_")
    
  # Create a data frame to hold all genes, for each trait
  all_genes <- data.frame(Gene = character(), Trait = character(), stringsAsFactors = FALSE)
  
  for (name in names(gene_tables)) {
    genes <- gene_tables[[name]]
    # Create a data frame with the trait name
    trait_genes <- data.frame(Gene = genes, Trait = name, stringsAsFactors = FALSE)
    all_genes <- rbind(all_genes, trait_genes)
  }
    
  all_genes <- all_genes %>%
    # Just keep the trait_name
    group_by(Trait) %>%
    mutate(Trait = strsplit(Trait, "_")[[1]][1:2] %>% paste(collapse = "_"))
  
  # Separate in distinctict dataframes by trait
  all_genes <- split(all_genes, all_genes$Trait)
  
  # Write to file
    for (trait in names(all_genes)) {
      trait_genes <- all_genes[[trait]] %>%
        distinct(Gene, Trait) %>%
        ungroup() %>%
        dplyr::select(Gene) %>%
        arrange(Gene)
      
      path <- file.path(gene_path, trait)
      createDir(path)
      write.table(trait_genes, file.path(path, paste0("all_genes_", trait, ".txt")), quote = FALSE, row.names = FALSE, col.names = FALSE, sep = "\t")
    }
  
  return(all_genes)
}

all_gene_tables_df <- all_gene_tables(gene_tables)

# Extract also significant bootstrap genes
gene_tables.boot <- lapply(names(caas_files_boot), function(name) get_gene_info(caas_files_boot[[name]], name, "boot"))
names(gene_tables.boot) <- names(caas_files_boot)
# Do the same for pat_1_dfs.boot, pat_2_dfs.boot, pat_3_dfs.boot, pat_12_dfs.boot, pat_13_dfs.boot
pat_1_gene_tables.boot <- lapply(names(pat_1_dfs.boot), function(name) get_gene_info(pat_1_dfs.boot[[name]], name, "pat1_boot"))
names(pat_1_gene_tables.boot) <- names(pat_1_dfs.boot)
pat_2_gene_tables.boot <- lapply(names(pat_2_dfs.boot), function(name) get_gene_info(pat_2_dfs.boot[[name]], name, "pat2_boot"))
names(pat_2_gene_tables.boot) <- names(pat_2_dfs.boot)
pat_3_gene_tables.boot <- lapply(names(pat_3_dfs.boot), function(name) get_gene_info(pat_3_dfs.boot[[name]], name, "pat3_boot"))
names(pat_3_gene_tables.boot) <- names(pat_3_dfs.boot)
pat_12_gene_tables.boot <- lapply(names(pat_12_dfs.boot), function(name) get_gene_info(pat_12_dfs.boot[[name]], name, "pat12_boot"))
names(pat_12_gene_tables.boot) <- names(pat_12_dfs.boot)
pat_13_gene_tables.boot <- lapply(names(pat_13_dfs.boot), function(name) get_gene_info(pat_13_dfs.boot[[name]], name, "pat13_boot"))
names(pat_13_gene_tables.boot) <- names(pat_13_dfs.boot)

# Merge all gene tables into a single data frame
all_gene_tables_boot <- function(gene_tables) {
  
  trait_name <- strsplit(names(gene_tables)[1], "_")[[1]][1:2] %>% paste(collapse = "_")
    
  # Create a data frame to hold all genes, for each trait
  all_genes <- data.frame(Gene = character(), Trait = character(), stringsAsFactors = FALSE)
  
  for (name in names(gene_tables)) {
    genes <- gene_tables[[name]]
    # Create a data frame with the trait name
    trait_genes <- data.frame(Gene = genes, Trait = name, stringsAsFactors = FALSE)
    all_genes <- rbind(all_genes, trait_genes)
  }
    
  all_genes <- all_genes %>%
    # Just keep the trait_name
    group_by(Trait) %>%
    mutate(Trait = strsplit(Trait, "_")[[1]][1:2] %>% paste(collapse = "_"))
    
  
  # Separate in distinctict dataframes by trait
  all_genes <- split(all_genes, all_genes$Trait)
  
  # Write to file
    for (trait in names(all_genes)) {
      trait_genes <- all_genes[[trait]]  %>%
        distinct(Gene, Trait) %>%
        ungroup() %>%
        dplyr::select(Gene) %>%
        arrange(Gene)
      
      path <- file.path(gene_path, trait)
      createDir(path)
      write.table(trait_genes, file.path(path, paste0("all_genes_", trait, "_boot.txt")), quote = FALSE, row.names = FALSE, col.names = FALSE, sep = "\t")
    }
  
  return(all_genes)
}

all_gene_tables_boot_df <- all_gene_tables_boot(gene_tables.boot)

```

## 6. Prepare the data for the agreggation step
In this section, we will prepare the data for the aggregation step. We will extract the commonalities between the nominally significant and non-nominally significant CAAS, and save the results to files.

```{r extract_commonalities_nominal}

# Apply to aggregated data
# Function to filter rows and create id2 column
filter_raw <- function(lst, pattern) {
  lst_filtered <- lst[grepl(pattern, names(lst))] %>%
    bind_rows(.id = "source")
  
  return(lst_filtered)
}

# Apply to aggregated data
agg_dfs.mp <- filter_raw(setup_dfs, "^malignant_")
agg_dfs.np <- filter_raw(setup_dfs, "^neoplasia_")

agg_dfs <- list(agg_dfs.mp, agg_dfs.np)
agg_dfs <- setNames(agg_dfs, c("malignant_prevalence", "neoplasia_prevalence"))

# Extract particular scenarios
# Create random string
myFun <- function(n = 100000) {
  a <- do.call(paste0, replicate(5, sample(LETTERS, n, TRUE), FALSE))
  paste0(a, sprintf("%04d", sample(9999, n, TRUE)), sample(LETTERS, n, TRUE))
}

# Check for duplicates
check_dups <- function(df) {
  dup_counts <- df %>%
    group_by(GenePos) %>%
    summarise(Count = n()) %>%
    filter(Count > 1)
  
  return(dup_counts)
}

malignant_dups <- check_dups(agg_dfs[[1]])
neoplasia_dups <- check_dups(agg_dfs[[2]])

# Create scenario 1, 2, 3, 12, 13 and 23
scenario_def <- function(df, name) {
  # Seed the random number generator
  set.seed(1998)
  
  df <- df %>%
    dplyr::mutate(
      AminoConv = Substitution
    ) %>%
    # Trait has values of the shape dfX_sp_pair_X.tab
    # Mutate to keep only the pair_X part
    dplyr::mutate(
      Trait = sub(".*_([^\\.]+).*", "\\1", Trait) # Keeps 'pair_X'
    ) %>%
    # Create index per row
    dplyr::mutate(Index = row_number()) %>%
    dplyr::group_by(Index) %>%
    dplyr::mutate(
      Tag = paste0("CAAS_", myFun(1))
    ) %>%
    dplyr::ungroup() %>%
    dplyr::select(-Index) %>%
    dplyr::rename(Contrast = Trait)
}

# Apply to aggregated data
agg_dfs.sc <- lapply(names(agg_dfs), function(name) scenario_def(agg_dfs[[name]], name))
names(agg_dfs.sc) <- names(agg_dfs)

extract_positions <- function(df) {
  # Create a grouped data frame to get unique positions for each pair
  pair_positions <- df %>%
    mutate(GenePos = paste(Gene, Position, sep = "_")) %>%
    group_by(GenePos) %>%
    summarise(Contrast = paste(unique(Contrast), collapse = ",")) %>%
    mutate(
      Unique = ifelse(grepl(",", Contrast), FALSE, TRUE) # Flag if the position belongs to multiple pairs
    )

  # Separate the data into two parts: unique and shared
  unique_positions <- pair_positions %>%
    filter(Unique == TRUE) %>%
    select(GenePos, Contrast)

  shared_positions <- pair_positions %>%
    filter(Unique == FALSE) %>%
    select(GenePos, Contrast)
  
  merged_positions <- df %>%
    mutate(GenePos = paste(Gene, Position, sep = "_")) %>%
    select(GenePos, Contrast, Tag)
  
  meta_caas <- df %>%
    mutate(GenePos = paste(Gene, Position, sep = "_")) %>%
    select(GenePos, Tag, Contrast, Pattern, AminoConv, Pvalue)
  
  return(list(unique_positions, shared_positions, merged_positions, meta_caas))
}

agg_positions <- lapply(agg_dfs.sc, extract_positions)
names(agg_positions) <- names(agg_dfs.sc)

subdf_names <- c("unique_positions", "shared_positions", "merged_positions", "meta_caas")


agg_positions <- lapply(agg_positions, function(x) {
  names(x) <- subdf_names
  return(x)
})

# Check for duplicates in meta_caas
metacaa_dups <- lapply(agg_positions, function(x) check_dups(x[[4]]))

# Extract the lists
extract_lists <- function(df, parent, name) {
  path <- file.path(paste0(caas_list_dir, "/4.Commonalities/not-significant/", parent))
  createDir(path)  # Ensure the directory exists
  write.table(df, file.path(path, paste0(name, ".output")), 
              quote = FALSE, row.names = FALSE, col.names = TRUE, sep = "\t")
  return(df)
}

# Iterate over the parent lists and extract each sublist
for (parent_name in names(agg_positions)) {
  sublists <- agg_positions[[parent_name]]
  
  for (sub_name in names(sublists)) {
    extract_lists(sublists[[sub_name]], parent_name, sub_name)
  }
}

# Apply to aggregated data
agg_dfs.mp <- filter_raw(setup_dfs.05_boot, "^malignant_")
agg_dfs.np <- filter_raw(setup_dfs.05_boot, "^neoplasia_")

agg_dfs.boot <- list(agg_dfs.mp, agg_dfs.np)
names(agg_dfs.boot) <- names(agg_dfs.sc)

# Apply to aggregated data
agg_dfs.sc.boot <- lapply(names(agg_dfs.boot), function(name) scenario_def(agg_dfs.boot[[name]], name))
names(agg_dfs.sc.boot) <- names(agg_dfs.boot)

# Extract particular scenarios
extract_positions <- function(df) {
  # Create a grouped data frame to get unique positions for each pair
  pair_positions <- df %>%
    mutate(GenePos = paste(Gene, Position, sep = "_")) %>%
    group_by(GenePos, Tag) %>%
    summarise(Contrast = paste(unique(Contrast), collapse = ",")) %>%
    mutate(
      Unique = ifelse(grepl(",", Contrast), FALSE, TRUE) # Flag if the position belongs to multiple pairs
    )

  # Separate the data into two parts: unique and shared
  unique_positions <- pair_positions %>%
    filter(Unique == TRUE) %>%
    select(GenePos, Contrast, Tag)
  
  shared_positions <- pair_positions %>%
    filter(Unique == FALSE) %>%
    select(GenePos, Contrast)
  
  merged_positions <- df %>%
    mutate(GenePos = paste(Gene, Position, sep = "_")) %>%
    group_by(GenePos) %>%
    select(GenePos, Contrast, Tag, Pvalue.boot) %>%
    dplyr::rename(Pvalue = Pvalue.boot)
  
  meta_caas <- df %>%
    mutate(GenePos = paste(Gene, Position, sep = "_")) %>%
    group_by(GenePos) %>%
    select(GenePos, Tag, Contrast, Pattern, AminoConv, Pvalue.boot) %>%
    dplyr::rename(Pvalue = Pvalue.boot)

  return(list(unique_positions, shared_positions, merged_positions, meta_caas))
}

agg_positions.boot <- lapply(agg_dfs.sc.boot, extract_positions)
names(agg_positions.boot) <- names(agg_dfs.sc.boot)

subdf_names <- c("unique_positions", "shared_positions", "merged_positions", "meta_caas")
# Set names for the subdataframes
agg_positions.boot <- lapply(agg_positions.boot, function(x) {
  names(x) <- subdf_names
  return(x)
})

# Extract the lists
extract_lists <- function(df, parent, name) {
  path <- file.path(paste0(caas_list_dir, "/4.Commonalities/significant/", parent))
  createDir(path)  # Ensure the directory exists
  write.table(df, file.path(path, paste0(name, ".output")), 
              quote = FALSE, row.names = FALSE, col.names = TRUE, sep = "\t")
  return(df)
}

# Iterate over the parent lists and extract each sublist
for (parent_name in names(agg_positions.boot)) {
  sublists <- agg_positions.boot[[parent_name]]
  
  for (sub_name in names(sublists)) {
    extract_lists(sublists[[sub_name]], parent_name, sub_name)
  }
}

# Now the idea here is just to highlight the significant positions in the non-significant dataframes
# For this we will use the meta_caas dataframes from the significant dataframes agg_positions.boot
# Note that both lists contain the same number of elements and the same names, only .boot is a subset of the other
# Structure: parent_list[[name]][["sublists"]]

highlight_significant_positions <- function(non_sig_list, sig_list) {
  # Iterate over parent elements (same names in both lists)
  result <- lapply(names(non_sig_list), function(parent_name) {
    
    non_sig_sublists <- non_sig_list[[parent_name]]
    sig_sublists <- sig_list[[parent_name]]
    
    # Iterate over sublists
    updated_sublists <- lapply(names(non_sig_sublists), function(sub_name) {
      non_sig_df <- non_sig_sublists[[sub_name]]
      sig_df <- sig_sublists[[sub_name]] %>%
        select(GenePos)

      print(head(non_sig_df))
      print(head(sig_df))
      # Check if dataframe is not empty
      if (nrow(non_sig_df) > 0 && nrow(sig_df) > 0) {
        # Flag rows as significant if they exist in sig_df
        non_sig_df$isSignificant <- non_sig_df$GenePos %in% sig_df$GenePoss
      } else {
        non_sig_df$isSignificant <- FALSE
      }
      
      return(non_sig_df)
    })

    names(updated_sublists) <- names(non_sig_sublists)  # Maintain sublist names
    return(updated_sublists)
  })

  names(result) <- names(non_sig_list)  # Maintain parent list names
  return(result)
}

# Apply function to flag significance in non-significant data
agg_positions.sig <- highlight_significant_positions(agg_positions, agg_positions.boot)

# Extract the lists
extract_lists <- function(df, parent, name) {
  path <- file.path(paste0(caas_list_dir, "/4.Commonalities/highlight/", parent))
  createDir(path)  # Ensure the directory exists
  write.table(df, file.path(path, paste0(name, ".output")), 
              quote = FALSE, row.names = FALSE, col.names = TRUE, sep = "\t")
  return(df)
}

# Iterate over the parent lists and extract each sublist
for (parent_name in names(agg_positions.sig)) {
  sublists <- agg_positions.sig[[parent_name]]
  
  for (sub_name in names(sublists)) {
    extract_lists(sublists[[sub_name]], parent_name, sub_name)
  }
}

# Check for duplicates in highlighted significant positions
highlighted_dups <- lapply(agg_positions.sig, function(x) check_dups(x[[4]]))

```
