==> Phylophere-integration/main_project_scripts/33.PGLS_analysis.Rmd <==
---
title: "PGLS analysis"
author: "Miguel Ramon"
date: "`r Sys.Date()`"
output:
html_document: default
pdf_document: default
---

################################## 
          
# PGLS

### PRIMATES' NEOPLASY PROJECT

### MIGUEL RAMON ALONSO

################################## 

## 1.Setup & Directory Definition

This section sets the working directory, loads common functions, and defines data and results folders.
```{r setup}

knitr::opts_knit$set(root.dir = "/home/miguel/IBE-UPF/PhD/NEOPLASY_PRIMATES")
getwd()

```

```{r dirdef, message=FALSE, warning=FALSE}

workingDir <- getwd()
dataDir <- file.path(workingDir, "Data")
resultsDir <- file.path(workingDir,"Malignancy_Primates/Out/")

# Load common objects, functions and libraries
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/libraries.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/commons.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/primate_functions.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/primate_objects.R"))
source(file.path(workingDir, "Malignancy_Primates/Scripts/functions/dunn_modified.R"))

```


## 2. Phylogenetic Signal Analysis

This section calculates the phylogenetic signal (Pagel’s lambda) for neoplasia prevalence, malignant prevalence, body mass, and maximum lifespan (MLS).

```{r phylo_sig, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}}
# Prepare cancer trait data

cancer_traits_

data_cancer <- cancer_traits_filt %>%
  dplyr::mutate(
    necropsy_count = adult_necropsy_count,
    NP = neoplasia_prevalence,
    NPper = neoplasia_prevalence * 100,
    MP = malignant_prevalence,
    MPper = malignant_prevalence * 100,
    BM_corr = log_body_mass,
    MLS_corr = log(MLS.anage)
  ) %>%
  dplyr::select(species, family, common_name, necropsy_count, NP, MP, BM_corr, MLS_corr)

# Check number of species
n <- nrow(data_cancer)
print(n)

# Extract traits and align with species
NP <- data_cancer$NP
names(NP) <- data_cancer$species
BM <- data_cancer$BM_corr
names(BM) <- data_cancer$species
MP <- data_cancer$MP
names(MP) <- data_cancer$species
MLS <- data_cancer$MLS_corr
names(MLS) <- data_cancer$species

# Align phylogenies with trait data
phylo_np <- geiger::treedata(phy, NP, warnings = FALSE)$phy
phylo_bm <- geiger::treedata(phy, BM, warnings = FALSE)$phy
phylo_mp <- geiger::treedata(phy, MP, warnings = FALSE)$phy
phylo_mls <- geiger::treedata(phy, MLS, warnings = FALSE)$phy

# Calculate phylogenetic signal (Pagel’s lambda)
lambda_np <- phylosig(phylo_np, NP, method = "lambda", test = TRUE)
lambda_mp <- phylosig(phylo_mp, MP, method = "lambda", test = TRUE)
lambda_bm <- phylosig(phylo_bm, BM, method = "lambda", test = TRUE)
lambda_mls <- phylosig(phylo_mls, MLS, method = "lambda", test = TRUE)

# Combine lambda values into a dataframe
lambda_values <- data.frame(
  trait = c("Neoplasia prevalence", "Malignant prevalence", "Body mass", "MLS"),
  lambda = c(lambda_np$lambda, lambda_mp$lambda, lambda_bm$lambda, lambda_mls$lambda),
  p_value = c(lambda_np$P, lambda_mp$P, lambda_bm$P, lambda_mls$P)
)

# Save lambda values
createDir(file.path(resultsDir, "1.Data-exploration/5.Phylogenetic_signal"))
write.csv(
  lambda_values,
  file.path(resultsDir, "1.Data-exploration/5.Phylogenetic_signal/lambda_values.csv"),
  row.names = FALSE
)
```


## 3. Peto’s Trends Analysis

This section examines family-level differences in neoplasia and malignant prevalence using linear models and pairwise comparisons.

```{r peto_trends, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}}
# Analyze neoplasia prevalence by family
mx_np <- lm(
  NP * 100 ~ family,
  data = data_cancer[data_cancer$family %in% names(table(data_cancer$family))[table(data_cancer$family) > 1], ]
)
newdata_np <- expand.grid(family = names(table(data_cancer$family))[table(data_cancer$family) > 1])
family_avg_np <- cbind(
  newdata_np,
  as.data.frame(emmeans::emmeans(mx_np, ~family)) %>% dplyr::select(-family)
)

# Perform pairwise comparisons for neoplasia prevalence
family_pairwise_np <- as.data.frame(emmeans::pairs(emmeans::emmeans(mx_np, ~family)))

# Analyze malignant prevalence by family
mx_mp <- lm(
  malignant_prevalence * 100 ~ family,
  data = data_cancer[data_cancer$family %in% names(table(data_cancer$family))[table(data_cancer$family) > 1], ]
)
newdata_mp <- expand.grid(family = names(table(data_cancer$family))[table(data_cancer$family) > 1])
family_avg_mp <- cbind(
  newdata_mp,
  as.data.frame(emmeans::emmeans(mx_mp, ~family)) %>% dplyr::select(-family)
)

# Perform pairwise comparisons for malignant prevalence
family_pairwise_mp <- as.data.frame(emmeans::pairs(emmeans::emmeans(mx_mp, ~family)))
```

## 4. PGLS Analysis: Body Mass and Neoplasia Prevalence

This section performs phylogenetic generalized least squares (PGLS) analyses to assess the relationship between body mass and neoplasia prevalence.

```{r peto_trend_bm_np, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}}
# Set row names for PGLS
row.names(data_cancer) <- data_cancer$species

# Define occurrence of neoplasia
data_cancer$occurrence_NP <- ifelse(data_cancer$NP == 0, 0, 1)
data_cancerNP <- data_cancer[as.character(phylo_np$tip.label), ]

# Prepare data for non-zero neoplasia prevalence
phy_nonzeroNP <- ape::drop.tip(phylo_np, which(!data_cancer$occurrence_NP == 1))
data_nonzeroNP <- data_cancer[data_cancer$species %in% phy_nonzeroNP$tip.label, ]

# PGLS model for body mass
bm_modNP <- ZIlogis(
  formbin = occurrence_NP ~ log(necropsy_count) + BM_corr,
  formlogis = car::logit(NP) ~ BM_corr,
  databin = data_cancer, phylobin = phylo_np,
  datalogis = data_nonzeroNP, phyloTree = phy_nonzeroNP, l0 = 1
)
sumZILogis(bm_modNP)

# Linear model for comparison
bm_lm <- lm(car::logit(NP) ~ BM_corr, data = data_cancer)
summary(bm_lm)

# PGLS model correcting for MLS
bm_modNP.mls <- ZIlogis(
  formbin = occurrence_NP ~ log(necropsy_count) + BM_corr + MLS_corr,
  formlogis = car::logit(NP) ~ BM_corr + MLS_corr,
  databin = data_cancer, phylobin = phylo_np,
  datalogis = data_nonzeroNP, phyloTree = phy_nonzeroNP, l0 = 1
)
sumZILogis(bm_modNP.mls)

# Linear model correcting for MLS
bm_lm.mls <- lm(car::logit(NP) ~ BM_corr + MLS_corr, data = data_cancer)
summary(bm_lm.mls)

# Save results to Excel
wb2 <- openxlsx::createWorkbook()
bm_dfs <- list(
  data.frame(sumZILogis(bm_modNP)),
  broom::tidy(summary(bm_lm)),
  data.frame(sumZILogis(bm_modNP.mls)),
  broom::tidy(summary(bm_lm.mls))
)
bm_sheetNames <- c("bm_pgls", "bm_lm", "bm_pgls_mls", "bm_lm_mls")

for (i in seq_along(bm_dfs)) {
  openxlsx::addWorksheet(wb2, bm_sheetNames[i])
  openxlsx::writeData(wb2, bm_sheetNames[i], bm_dfs[[i]])
}

openxlsx::saveWorkbook(
  wb2,
  file.path(resultsDir, "1.Data-exploration/5.Phylogenetic_signal/bm_pgls_lm_np.xlsx"),
  overwrite = TRUE
)
```

## 5. PGLS Analysis: MLS and Neoplasia Prevalence

This section performs PGLS analyses to assess the relationship between maximum lifespan (MLS) and neoplasia prevalence.

```{r peto_trend_mls_np, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}}
# PGLS model for MLS
mls_modNP <- ZIlogis(
  formbin = occurrence_NP ~ log(necropsy_count) + MLS_corr,
  formlogis = car::logit(NP) ~ MLS_corr,
  databin = data_cancer, phylobin = phylo_np,
  datalogis = data_nonzeroNP, phyloTree = phy_nonzeroNP, l0 = 1
)
sumZILogis(mls_modNP)

# Linear model for comparison
mls_lm <- lm(car::logit(NP) ~ MLS_corr, data = data_cancer)
summary(mls_lm)

# PGLS model correcting for body mass
mls_modNP.bm <- ZIlogis(
  formbin = occurrence_NP ~ log(necropsy_count) + MLS_corr + BM_corr,
  formlogis = car::logit(NP) ~ MLS_corr + BM_corr,
  databin = data_cancer, phylobin = phylo_np,
  datalogis = data_nonzeroNP, phyloTree = phy_nonzeroNP, l0 = 1
)
sumZILogis(mls_modNP.bm)

# Linear model correcting for body mass
mls_lm.bm <- lm(car::logit(NP) ~ MLS_corr + BM_corr, data = data_cancer)
summary(mls_lm.bm)

# Save results to Excel
wb3 <- openxlsx::createWorkbook()
mls_dfs <- list(
  data.frame(sumZILogis(mls_modNP)),
  broom::tidy(summary(mls_lm)),
  data.frame(sumZILogis(mls_modNP.bm)),
  broom::tidy(summary(mls_lm.bm))
)
mls_sheetNames <- c("mls_pgls", "mls_lm", "mls_pgls_bm", "mls_lm_bm")

for (i in seq_along(mls_dfs)) {
  openxlsx::addWorksheet(wb3, mls_sheetNames[i])
  openxlsx::writeData(wb3, mls_sheetNames[i], mls_dfs[[i]])
}

openxlsx::saveWorkbook(
  wb3,
  file.path(resultsDir, "1.Data-exploration/5.Phylogenetic_signal/mls_pgls_lm_np.xlsx"),
  overwrite = TRUE
)
```

## 6. PGLS Analysis: Body Mass and Malignant Prevalence

This section performs PGLS analyses to assess the relationship between body mass and malignant prevalence.

```{r peto_trend_bm_mp, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}}
# Define occurrence of malignancy
data_cancer$occurrence_MP <- ifelse(data_cancer$MP == 0, 0, 1)
data_cancerMP <- data_cancer[as.character(phylo_mp$tip.label), ]

# Prepare data for non-zero malignant prevalence
phy_nonzeroMP <- ape::drop.tip(phylo_mp, which(!data_cancer$occurrence_MP == 1))
data_nonzeroMP <- data_cancer[data_cancer$species %in% phy_nonzeroMP$tip.label, ]

# PGLS model for body mass
bm_modMP <- ZIlogis(
  formbin = occurrence_MP ~ log(necropsy_count) + BM_corr,
  formlogis = car::logit(MP) ~ BM_corr,
  databin = data_cancer, phylobin = phylo_mp,
  datalogis = data_nonzeroMP, phyloTree = phy_nonzeroMP, l0 = 1
)
sumZILogis(bm_modMP)

# Linear model for comparison
bm_lm <- lm(car::logit(MP) ~ BM_corr, data = data_cancer)
summary(bm_lm)

# PGLS model correcting for MLS
bm_modMP.mls <- ZIlogis(
  formbin = occurrence_MP ~ log(necropsy_count) + BM_corr + MLS_corr,
  formlogis = car::logit(MP) ~ BM_corr + MLS_corr,
  databin = data_cancer, phylobin = phylo_mp,
  datalogis = data_nonzeroMP, phyloTree = phy_nonzeroMP, l0 = 1
)
sumZILogis(bm_modMP.mls)

# Linear model correcting for MLS
bm_lm.mls <- lm(car::logit(MP) ~ BM_corr + MLS_corr, data = data_cancer)
summary(bm_lm.mls)

# Save results to Excel
wb4 <- openxlsx::createWorkbook()
bm_dfs <- list(
  data.frame(sumZILogis(bm_modMP)),
  broom::tidy(summary(bm_lm)),
  data.frame(sumZILogis(bm_modMP.mls)),
  broom::tidy(summary(bm_lm.mls))
)
bm_sheetNames <- c("bm_pgls", "bm_lm", "bm_pgls_mls", "bm_lm_mls")

for (i in seq_along(bm_dfs)) {
  openxlsx::addWorksheet(wb4, bm_sheetNames[i])
  openxlsx::writeData(wb4, bm_sheetNames[i], bm_dfs[[i]])
}

openxlsx::saveWorkbook(
  wb4,
  file.path(resultsDir, "1.Data-exploration/5.Phylogenetic_signal/bm_pgls_lm_mp.xlsx"),
  overwrite = TRUE
)
```

## 7. PGLS Analysis: MLS and Malignant Prevalence

This section performs PGLS analyses to assess the relationship between maximum lifespan (MLS) and malignant prevalence.

```{r peto_trend_mls_mp, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}}
# PGLS model for MLS
mls_modMP <- ZIlogis(
  formbin = occurrence_MP ~ log(necropsy_count) + MLS_corr,
  formlogis = car::logit(MP) ~ MLS_corr,
  databin = data_cancer, phylobin = phylo_mp,
  datalogis = data_nonzeroMP, phyloTree = phy_nonzeroMP, l0 = 1
)
sumZILogis(mls_modMP)

# Linear model for comparison
mls_lm <- lm(car::logit(MP) ~ MLS_corr, data = data_cancer)
summary(mls_lm)

# PGLS model correcting for body mass
mls_modMP.bm <- ZIlogis(
  formbin = occurrence_MP ~ log(necropsy_count) + MLS_corr + BM_corr,
  formlogis = car::logit(MP) ~ MLS_corr + BM_corr,
  databin = data_cancer, phylobin = phylo_mp,
  datalogis = data_nonzeroMP, phyloTree = phy_nonzeroMP, l0 = 1
)
sumZILogis(mls_modMP.bm)

# Linear model correcting for body mass
mls_lm.bm <- lm(car::logit(MP) ~ MLS_corr + BM_corr, data = data_cancer)
summary(mls_lm.bm)

# Save results to Excel
wb5 <- openxlsx::createWorkbook()
mls_dfs <- list(
  data.frame(sumZILogis(mls_modMP)),
  broom::tidy(summary(mls_lm)),
  data.frame(sumZILogis(mls_modMP.bm)),
  broom::tidy(summary(mls_lm.bm))
)
mls_sheetNames <- c("mls_pgls", "mls_lm", "mls_pgls_bm", "mls_lm_bm")

for (i in seq_along(mls_dfs)) {
  openxlsx::addWorksheet(wb5, mls_sheetNames[i])
  openxlsx::writeData(wb5, mls_sheetNames[i], mls_dfs[[i]])
}

openxlsx::saveWorkbook(
  wb5,
  file.path(resultsDir, "1.Data-exploration/5.Phylogenetic_signal/mls_pgls_lm_mp.xlsx"),
  overwrite = TRUE
)
```

## 8. PGLS Analysis: Neoplasia vs. Malignant Prevalence

This section performs PGLS analyses to assess the relationship between neoplasia prevalence and malignant prevalence.

```{r peto_trend_np_mp, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}}
# PGLS model for neoplasia vs. malignant prevalence
np_modMP <- ZIlogis(
  formbin = occurrence_MP ~ log(necropsy_count) + car::logit(NP),
  formlogis = car::logit(MP) ~ car::logit(NP),
  databin = data_cancer, phylobin = phylo_mp,
  datalogis = data_nonzeroMP, phyloTree = phy_nonzeroMP, l0 = 1
)
sumZILogis(np_modMP)

# Linear model for comparison
np_lm <- lm(car::logit(MP) ~ car::logit(NP), data = data_cancer)
summary(np_lm)

# PGLS model correcting for MLS
np_modMP.mls <- ZIlogis(
  formbin = occurrence_MP ~ log(necropsy_count) + car::logit(NP) + MLS_corr,
  formlogis = car::logit(MP) ~ car::logit(NP) + MLS_corr,
  databin = data_cancer, phylobin = phylo_mp,
  datalogis = data_nonzeroMP, phyloTree = phy_nonzeroMP, l0 = 1
)
sumZILogis(np_modMP.mls)

# Linear model correcting for MLS
np_lm.mls <- lm(car::logit(MP) ~ car::logit(NP) + MLS_corr, data = data_cancer)
summary(np_lm.mls)

# Save results to Excel
wb6 <- openxlsx::createWorkbook()
np_mp_dfs <- list(
  data.frame(sumZILogis(np_modMP)),
  broom::tidy(summary(np_lm)),
  data.frame(sumZILogis(np_modMP.mls)),
  broom::tidy(summary(np_lm.mls))
)
np_mp_sheetNames <- c("np_mp", "np_lm", "np_mp_mls", "np_lm_mls")

for (i in seq_along(np_mp_dfs)) {
  openxlsx::addWorksheet(wb6, np_mp_sheetNames[i])
  openxlsx::writeData(wb6, np_mp_sheetNames[i], np_mp_dfs[[i]])
}

openxlsx::saveWorkbook(
  wb6,
  file.path(resultsDir, "1.Data-exploration/5.Phylogenetic_signal/mls_pgls_lm_npmp.xlsx"),
  overwrite = TRUE
)
```

## 9. PGLS Visualization

This section generates scatter plots to visualize PGLS relationships between traits, including regression lines and phylogenetic signal annotations.

```{r peto_plots, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}
# Function to generate PGLS scatter plots
peto_plots <- function(data, response, predictor, name, model) {
  filter_trait <- data %>%
    dplyr::filter(!is.na(!!sym(predictor)))
  
  t1 <- filter_trait %>% dplyr::pull(!!sym(response)) * 100
  t2 <- filter_trait %>% dplyr::pull(!!sym(predictor))
  common_name <- filter_trait %>% dplyr::pull(common_name)
  necropsy_count <- filter_trait %>% dplyr::pull(necropsy_count)
  species <- filter_trait %>% dplyr::pull(species)
  family <- filter_trait %>% dplyr::pull(family)
  
  merged_t <- data.frame(t1, t2, necropsy_count, species, family)
  rownames(merged_t) <- merged_t$species
  
  # Extract regression coefficients
  intercept <- model$coefficients[1]
  slope <- model$coefficients[2]
  
  # Define x and y ranges
  x_min <- min(merged_t$t2, na.rm = TRUE)
  x_max <- max(merged_t$t2, na.rm = TRUE)
  y_min <- 1 / (1 + exp(-(intercept + slope * x_min))) * 100
  y_max <- 1 / (1 + exp(-(intercept + slope * x_max))) * 100
  
  # Extract lambda and p-value
  lambda <- summary(model)$modelStruct$corStruct[1]
  pval <- summary(model)$tTable[2, 4]
  
  # Create scatter plot
  ggplot2::ggplot(merged_t, aes(x = t2, y = t1, color = family)) +
    ggplot2::geom_point(size = log(necropsy_count * 2.5), alpha = 0.5) +
    ggplot2::scale_color_manual(values = dark_family_palette, breaks = NULL) +
    ggrepel::geom_text_repel(
      aes(label = common_name), hjust = 0, vjust = 0,
      size = log(necropsy_count * 0.5), family = "Inter", max.overlaps = 50,
      position = ggrepel::position_nudge_repel(x = 0.005, y = 0.01),
      force = 0.1, direction = "both", box.padding = 0.5, seed = 1998,
      alpha = 1, show.legend = FALSE
    ) +
    ggplot2::geom_segment(
      aes(x = x_min, y = y_min, xend = x_max, yend = y_max),
      color = "black", linetype = "dashed"
    ) +
    ggplot2::geom_smooth(method = "lm", color = "black", se = FALSE) +
    ggplot2::labs(
      x = name,
      y = paste0(response, " (%)"),
      caption = paste0(
        "Pagel's lambda: ", round(lambda, 2),
        "\nP-value: ", round(pval, 4)
      )
    ) +
    ggplot2::theme_minimal() +
    ggplot2::theme(
      axis.title.x = ggplot2::element_text(size = 16, margin = ggplot2::margin(t = 20)),
      axis.title.y = ggplot2::element_text(size = 16, margin = ggplot2::margin(r = 20)),
      plot.caption = ggplot2::element_text(size = 16),
      legend.position = "none"
    )
}

# Generate plots
bm_plot_np <- peto_plots(data_cancerNP, "NP", "BM_corr", "Body Mass (log.kg)", bm_modNP$wlogis)
bm_plot_mp <- peto_plots(data_cancerMP, "MP", "BM_corr", "Body Mass (log.kg)", bm_modMP$wlogis)
mls_plot_np <- peto_plots(data_cancerNP, "NP", "MLS_corr", "MLS", mls_modNP$wlogis)
mls_plot_mp <- peto_plots(data_cancerMP, "MP", "MLS_corr", "MLS", mls_modMP$wlogis)

# Display plots
print(bm_plot_np)
print(bm_plot_mp)
print(mls_plot_np)
print(mls_plot_mp)

# Save plots
plot_dir <- file.path(resultsDir, "1.Data-exploration/6.Correlations")
createDir(plot_dir)
ggplot2::ggsave(
  plot = bm_plot_np, filename = file.path(plot_dir, "bm_plot_np.png"),
  dpi = "retina", width = 10, height = 10
)
ggplot2::ggsave(
  plot = bm_plot_mp, filename = file.path(plot_dir, "bm_plot_mp.png"),
  dpi = "retina", width = 10, height = 10
)
ggplot2::ggsave(
  plot = mls_plot_np, filename = file.path(plot_dir, "mls_plot_np.png"),
  dpi = "retina", width = 10, height = 10
)
ggplot2::ggsave(
  plot = mls_plot_mp, filename = file.path(plot_dir, "mls_plot_mp.png"),
  dpi = "retina", width = 10, height = 10
)
```

## 10. PGLS Visualization: Neoplasia vs. Malignant Prevalence

This section generates a scatter plot specifically for the relationship between neoplasia and malignant prevalence.

```{r peto_plots_print, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}
# Function to generate PGLS scatter plot for NP vs. MP
peto_plots <- function(data, response, predictor, name, model) {
  filter_trait <- data %>%
    dplyr::filter(!is.na(!!sym(predictor)))
  
  t1 <- filter_trait %>% dplyr::pull(!!sym(response)) * 100
  t2 <- filter_trait %>% dplyr::pull(!!sym(predictor))
  t2_mod <- car::logit(t2)
  t2 <- t2 * 100
  common_name <- filter_trait %>% dplyr::pull(common_name)
  necropsy_count <- filter_trait %>% dplyr::pull(necropsy_count)
  species <- filter_trait %>% dplyr::pull(species)
  family <- filter_trait %>% dplyr::pull(family)
  
  merged_t <- data.frame(t1, t2, t2_mod, necropsy_count, species, family)
  rownames(merged_t) <- merged_t$species
  
  # Extract regression coefficients
  intercept <- model$coefficients[1]
  slope <- model$coefficients[2]
  
  # Define x and y ranges
  x_min_mod <- min(merged_t$t2_mod, na.rm = TRUE)
  x_max_mod <- max(merged_t$t2_mod, na.rm = TRUE)
  x_min <- min(merged_t$t2, na.rm = TRUE)
  x_max <- max(merged_t$t2, na.rm = TRUE)
  y_min <- 1 / (1 + exp(-(intercept + slope * x_min_mod))) * 100
  y_max <- 1 / (1 + exp(-(intercept + slope * x_max_mod))) * 100
  
  # Extract lambda and p-value
  lambda <- summary(model)$modelStruct$corStruct[1]
  pval <- summary(model)$tTable[2, 4]
  
  # Create scatter plot
  ggplot2::ggplot(merged_t, aes(x = t2, y = t1, color = family)) +
    ggplot2::geom_point(size = log(necropsy_count * 2.5), alpha = 0.5) +
    ggplot2::scale_color_manual(values = dark_family_palette, breaks = NULL) +
    ggrepel::geom_text_repel(
      aes(label = common_name), hjust = 0, vjust = 0,
      size = log(necropsy_count * 0.5), family = "Inter", max.overlaps = 50,
      position = ggrepel::position_nudge_repel(x = 0.005, y = 0.01),
      force = 0.1, direction = "both", box.padding = 0.5, seed = 1998,
      alpha = 1, show.legend = FALSE
    ) +
    ggplot2::geom_segment(
      aes(x = x_min, y = y_min, xend = x_max, yend = y_max),
      color = "black", linetype = "dashed"
    ) +
    ggplot2::geom_smooth(method = "lm", color = "black", se = FALSE) +
    ggplot2::labs(
      x = name,
      y = "Neoplasia prevalence (%)",
      caption = paste0(
        "Pagel's lambda: ", round(lambda, 2),
        "\nP-value: ", round(pval, 8)
      )
    ) +
    ggplot2::theme_minimal() +
    ggplot2::theme(
      axis.title.x = ggplot2::element_text(size = 16, margin = ggplot2::margin(t = 20)),
      axis.title.y = ggplot2::element_text(size = 16, margin = ggplot2::margin(r = 20)),
      plot.caption = ggplot2::element_text(size = 16),
      legend.position = "none"
    )
}

# Generate and save plot
np_plot_mp <- peto_plots(data_cancerMP, "NP", "MP", "Malignant prevalence (%)", np_modMP$wlogis)
ggplot2::ggsave(
  plot = np_plot_mp, filename = file.path(plot_dir, "np_plot_mp.png"),
  dpi = "retina", width = 10, height = 10
)

# Display plot
print(np_plot_mp)
```

## 11. Residual Scatter Plots

This section generates scatter plots of residuals from PGLS models to assess model fit.

```{r residual_plots, fig.height=10, fig.width=15, message=FALSE, warning=FALSE, dev='png'}
# Function to generate residual scatter plots
peto_plots <- function(data, response, predictor, name, model) {
  filter_trait <- data %>%
    dplyr::filter(!is.na(!!sym(predictor)))
  
  t2 <- model$fitted
  t2_names <- names(t2)
  filter_trait <- filter_trait %>% dplyr::filter(species %in% t2_names)
  
  t1 <- model$residuals
  t1_names <- names(t1)
  filter_trait <- filter_trait %>% dplyr::filter(species %in% t1_names)
  
  common_name <- filter_trait %>% dplyr::pull(common_name)
  necropsy_count <- filter_trait %>% dplyr::pull(necropsy_count)
  species <- filter_trait %>% dplyr::pull(species)
  family <- filter_trait %>% dplyr::pull(family)
  
  merged_t <- data.frame(t1, t2, necropsy_count, species, family)
  rownames(merged_t) <- merged_t$species
  
  # Extract regression coefficients
  intercept <- model$coefficients[1]
  slope <- model$coefficients[2]
  
  # Define x and y ranges
  x_min <- min(merged_t$t2, na.rm = TRUE)
  x_max <- max(merged_t$t2, na.rm = TRUE)
  y_min <- 1 / (1 + exp(-(intercept + slope * x_min))) * 100
  y_max <- 1 / (1 + exp(-(intercept + slope * x_max))) * 100
  
  # Extract lambda and p-value
  lambda <- summary(model)$modelStruct$corStruct[1]
  pval <- summary(model)$tTable[2, 4]
  
  # Create residual scatter plot
  ggplot2::ggplot(merged_t, aes(x = t2, y = t1, color = family)) +
    ggplot2::geom_point(size = log(necropsy_count * 2.5), alpha = 0.5) +
    ggplot2::scale_color_manual(values = dark_family_palette, breaks = NULL) +
    ggrepel::geom_text_repel(
      aes(label = common_name), hjust = 0, vjust = 0,
      size = log(necropsy_count * 0.5), family = "Inter", max.overlaps = 50,
      position = ggrepel::position_nudge_repel(x = 0.005, y = 0.01),
      force = 0.1, direction = "both", box.padding = 0.5, seed = 1998,
      alpha = 1, show.legend = FALSE
    ) +
    ggplot2::geom_smooth(method = "lm", color = "black", se = FALSE) +
    ggplot2::labs(
      x = "Fitted values",
      y = "Residuals",
      caption = paste0(
        "Pagel's lambda: ", round(lambda, 2),
        "\nP-value: ", round(pval, 8)
      )
    ) +
    ggplot2::theme_minimal() +
    ggplot2::theme(
      axis.title.x = ggplot2::element_text(size = 16, margin = ggplot2::margin(t = 20)),
      axis.title.y = ggplot2::element_text(size = 16, margin = ggplot2::margin(r = 20)),
      plot.caption = ggplot2::element_text(size = 16),
      legend.position = "none"
    )
}

# Generate residual plots
mls_bmresiduals_mp <- peto_plots(data_cancerMP, "MP", "MLS", "MP~MLS+BM Residuals", mls_modMP.bm$wlogis)
mls_bmresiduals_np <- peto_plots(data_cancerMP, "NP", "MLS", "NP~MLS+BM Residuals", mls_modNP.bm$wlogis)
np_mlsresiduals_mp <- peto_plots(data_cancerMP, "MP", "NP", "MP~NP+MLS Residuals", np_modMP.mls$wlogis)

# Display plots
print(np_mlsresiduals_mp)
print(mls_bmresiduals_np)
print(mls_bmresiduals_mp)

# Save plots
ggplot2::ggsave(
  plot = mls_bmresiduals_mp, filename = file.path(plot_dir, "mls_bmresiduals_mp.png"),
  dpi = "retina", width = 10, height = 10
)
ggplot2::ggsave(
  plot = mls_bmresiduals_np, filename = file.path(plot_dir, "mls_bmresiduals_np.png"),
  dpi = "retina", width = 10, height = 10
)
ggplot2::ggsave(
  plot = np_mlsresiduals_mp, filename = file.path(plot_dir, "np_mlsresiduals_mp.png"),
  dpi = "retina", width = 10, height = 10
)
```
